refactor(eclipse): retire fork driver; wasm backend becomes engine_tex
Delete the forked TeX engine driver and make the in-process wasm engine the
sole TeX engine, keeping the engine_tex name for the role (it still loads
.fmt formats — the wasm-ness is an implementation detail):

- engine_wasm.c -> engine_tex.c, engine_wasm.h -> engine_tex.h,
  engine_wasm_<eng>.c -> engine_tex_<eng>.c
- txp_create_wasm_engine -> txp_create_tex_engine (engine.h now declares the
  4-arg wasm constructor; the 7-arg fork constructor is gone)
- main.c calls txp_create_tex_engine directly (no TEXPRESSO_FORK fallback)
- Makefile: engine_tex + profiles now require WASM_ENGINE_DIR

The fork's fence/rollback algorithm is preserved verbatim in fence_ref.c
(not compiled) for the later fine-grained port. The wasm runtime
(src/engine-wasm/wasm_host.*, wasm2c) keeps its names. sprotocol, the editor
protocol, and the pdf/dvi viewers are untouched. NOTE: this removes only the
frontend fork driver, not the engine source (src/engine).
