+++ src/frontend/main.c
@@ -1233,20 +1233,8 @@ bool texpresso_main(struct persistent_state *ps)
 
     if (doc_ext && (strcmp(doc_ext, "dvi") == 0 || strcmp(doc_ext, "xdv") == 0))
       ui->eng = txp_create_dvi_engine(ps->ctx, ps->doc_name, hooks);
-#ifdef HAVE_WASM_ENGINE
-    /* The in-process wasm engine is the default TeX engine. The forked engine is
-     * retained only as a transitional fallback (TEXPRESSO_FORK) until it is
-     * removed. */
-    else if (!getenv("TEXPRESSO_FORK"))
-    {
-      fprintf(stderr, "[info] using in-process wasm engine\n");
-      ui->eng = txp_create_wasm_engine(ps->ctx, engine_path, ps->doc_name, hooks);
-    }
-#endif
-    else
-      ui->eng = txp_create_tex_engine(ps->ctx, engine_path, using_texlive,
-                                      ps->stream_mode, ps->inclusion_path,
-                                      ps->doc_name, hooks);
+    else /* .tex: the in-process wasm TeX engine */
+      ui->eng = txp_create_tex_engine(ps->ctx, engine_path, ps->doc_name, hooks);
   }
 
   ui->sdl_renderer = ps->renderer;
